+++ axp192.h
@@ -185,7 +185,7 @@ typedef int32_t axp192_err_t;
 
 axp192_err_t axp192_init(const axp192_t *axp);
 axp192_err_t axp192_read(const axp192_t *axp, uint8_t reg, void *buffer);
-axp192_err_t axp192_ioctl(const axp192_t *axp, uint16_t command, ...);
+axp192_err_t axp192_ioctl(const axp192_t *axp, int command, ...);
 
 #ifdef __cplusplus
 }
